Re-write Download_file.py to handle input errors

+++ src/Download_file.py
@@ -11,6 +11,8 @@ Options:
 
 from docopt import docopt
 import requests
+from os import path
+import sys
 
 opt = docopt(__doc__)
 
@@ -19,10 +21,20 @@ def main(opt):
     source = opt["--source"]
     destination = opt["--destination"]
 
-    req = requests.get(source)
+    if path.exists(destination) == False:
+        raise NameError("The path to destination file does not exist.")
+
+    try:
+        req = requests.get(source)
+    except:
+        print(
+            "Failed to connect to the online source file. Please check whether it exists."
+        )
+        sys.exit(1)
+
+    # req = requests.get(source)
     url_content = req.content
     csv_file = open(destination, "wb")
-
     csv_file.write(url_content)
     csv_file.close()
